[dev]front: create context to handle homepage sorting
From my mind, that's not HomeView's job to handle sort logic, that have to be moved in his own files to be reusable anywhere. Context is perfect for that! In addition to that, to handle state manipulation cleanly, I connected a reducer to our new context.
To conclude, I prepare some util functions which will be used by HomeView component to populate future select component.